@@ -1,71 +1,69 @@
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[project]
+[tool.poetry]
 name = "mdvtools"
 version = "0.0.1"
-authors = [
-    { name= "Martin Sergeant", email="m.j.sergeant1@googlemail.com"},
-    { name= "PeterTodd", email="peter.todd@well.ox.ac.uk" },
-]
 description = "Multi Dimensional Viewer (MDV) is web based application for analyzing data"
-readme = "readme.md"
-requires-python = ">=3.7"
+authors = [
+    "Martin Sergeant <m.j.sergeant1@googlemail.com>",
+    "PeterTodd <peter.todd@well.ox.ac.uk>",
+]
+readme = "README.md"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
     "Operating System :: OS Independent",
 ]
-dependencies = [
- "anndata==0.9.1",
- "click==8.1.3",
- "colorama==0.4.6",
- "contourpy==1.1.1",
- "cycler==0.12.0",
- "fasteners==0.18",
- "Flask==2.2.5",
- "Flask-SocketIO==5.3.4",
- "fonttools==4.43.0",
- "h5py==3.10.0",
- "itsdangerous==2.1.2",
- "Jinja2==3.1.2",
- "joblib==1.2.0",
- "jupyterlab==4.1.4",
- "kiwisolver==1.4.5",
- "llvmlite==0.42.0",
- "MarkupSafe==2.1.2",
- "matplotlib==3.8.0",
- "mudata==0.2.3",
- "natsort==8.3.1",
- "networkx==3.1",
- "numba==0.59.0",
- "numpy==1.26.4",
- "ome-types==0.5.0",
- "packaging==23.2",
- "pandas==2.1.1",
- "patsy==0.5.3",
- "Pillow==10.0.1",
- "pynndescent==0.5.10",
- "pyparsing==3.1.1",
- "python-dateutil==2.8.2",
- "pytz==2023.3.post1",
- "setuptools==69.1.1",
- "scanpy==1.9.8",
- "scikit-learn==1.3.0",
- "scipy==1.11.3",
- "seaborn==0.13.0",
- "session-info==1.0.0",
- "six==1.16.0",
- "statsmodels==0.14.0",
- "stdlib-list==0.8.0",
- "threadpoolctl==3.1.0",
- "tqdm==4.65.0",
- "tzdata==2023.3",
- "umap-learn==0.5.3",
- "Werkzeug==2.2.3"
-]
 
-[project.urls]
-"Homepage" = "https://github.com/Taylor-CCB-Group/MDV"
-"Bug Tracker" = "https://github.com/Taylor-CCB-Group/MDV/issues"
+
+[tool.poetry.dependencies]
+python = "^3.12,<3.13"
+anndata = "^0.9.1"
+fasteners = "^0.18"
+Flask = "^2.2.5"
+Flask-SocketIO = "^5.3.4"
+h5py = "^3.10.0"
+mudata = "^0.2.3"
+numpy = "^1.26.4"
+pandas = "^2.1.1"
+scanpy = "^1.9.8"
+scipy = "^1.11.3"
+werkzeug = "^3.0.2"
+
+
+[tool.poetry.group.dev.dependencies]
+pyright = "^1.1.358"
+pytest = "^8.1.1"
+ruff = "^0.3.7"
+ome-types = "^0.5.0"
+jupyterlab = "^4.1.8"
+leidenalg = "^0.10.2"
+matplotlib = "^3.8.4"
+setuptools = "^69.5.1"
+packaging = "^24.0"
+pillow = "^10.3.0"
+seaborn = "^0.13.2"
+umap-learn = "^0.5.6"
+statsmodels = "^0.14.2"
+
+
+
+[tool.poetry.group.docs.dependencies]
+sphinx-rtd-theme = "^2.0.0"
+sphinx = "^7.3.7"
+sphinx-autoapi = "^3.0.0"
+
+
+[tool.poetry.group.backend.dependencies]
+flask-sqlalchemy = "^3.1.1"
+
+[tool.pyright]
+venvPath = "../venv/"
+typeCheckingMode = "standard"
+
+# [project.urls]
+# "Homepage" = "https://github.com/Taylor-CCB-Group/MDV"
+# "Bug Tracker" = "https://github.com/Taylor-CCB-Group/MDV/issues"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
