@@ -14,20 +14,19 @@
 #
 import os
 import sys
-import sphinx_rtd_theme
-sys.path.insert(0, os.path.abspath('../'))
 
+sys.path.insert(0, os.path.abspath("../"))
 
 # -- Project information -----------------------------------------------------
 
-project = 'MDV'
-copyright = '2021, Martin Sergeant'
-author = 'Martin Sergeant'
+project = "MDV"
+copyright = "2021, Martin Sergeant"
+author = "Martin Sergeant"
 
 # The short X.Y version
-version = ''
+version = ""
 # The full version, including alpha/beta/rc tags
-release = ''
+release = ""
 
 
 # -- General configuration ---------------------------------------------------
@@ -40,47 +39,47 @@ release = ''
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-     'sphinx.ext.viewcode',
-	'sphinx.ext.autodoc',
-	'autoapi.extension',
-	'sphinx.ext.napoleon',
-	'sphinx_rtd_theme',
-    "sphinx_js"
+    "sphinx.ext.viewcode",
+    "sphinx.ext.autodoc",
+    "autoapi.extension",
+    "sphinx.ext.napoleon",
+    "sphinx_rtd_theme",
+    "sphinx_js",
 ]
-autoapi_type = 'python' #i think this is the default and only values allowed
-autoapi_dirs = ['../../python/mdvtools']
-#required for vscode to fin the autoapi files and allows linting/on the fly builds
+autoapi_type = "python"  # i think this is the default and only values allowed
+autoapi_dirs = ["../../python/mdvtools"]
+# required for vscode to fin the autoapi files and allows linting/on the fly builds
 autoapi_keep_files = True
-#don't build all the api docs - spcificc ones will be added manually
+# don't build all the api docs - spcificc ones will be added manually
 autoapi_generate_api_docs = True
 autoapi_add_toctree_entry = False
 
 
 html_logo = "../../images/mdv_logo.png"
 html_theme_options = {
-    'logo_only': True,
+    "logo_only": True,
     #'display_version': False,
 }
 
 # TBD:
 # js_language = 'typescript'
 
-#is there a simpler way to sepcify than having to list every folder?
-#js_source_path = ['../../src/charts','../../src/dataloaders','../../src/utilities']
-js_source_path=['../../src/datastore','../../src/charts','../../src/dataloaders']
-root_for_relative_js_paths = '../../src'
+# is there a simpler way to sepcify than having to list every folder?
+# js_source_path = ['../../src/charts','../../src/dataloaders','../../src/utilities']
+js_source_path = ["../../src/datastore", "../../src/charts", "../../src/dataloaders"]
+root_for_relative_js_paths = "../../src"
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
 #
 # source_suffix = ['.rst', '.md']
-source_suffix = '.rst'
+source_suffix = ".rst"
 
 # The master toctree document.
-master_doc = 'index'
+master_doc = "index"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
@@ -92,7 +91,7 @@ language = "en"
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 # The name of the Pygments (syntax highlighting) style to use.
 pygments_style = None
@@ -103,7 +102,7 @@ pygments_style = None
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'sphinx_rtd_theme'
+html_theme = "sphinx_rtd_theme"
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
@@ -114,9 +113,7 @@ html_theme = 'sphinx_rtd_theme'
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
-
-
+html_static_path = ["_static"]
 
 
 # Custom sidebar templates, must be a dictionary that maps document names
@@ -133,7 +130,7 @@ html_static_path = ['_static']
 # -- Options for HTMLHelp output ---------------------------------------------
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'MDVdoc'
+htmlhelp_basename = "MDVdoc"
 
 
 # -- Options for LaTeX output ------------------------------------------------
@@ -142,15 +139,12 @@ latex_elements = {
     # The paper size ('letterpaper' or 'a4paper').
     #
     # 'papersize': 'letterpaper',
-
     # The font size ('10pt', '11pt' or '12pt').
     #
     # 'pointsize': '10pt',
-
     # Additional stuff for the LaTeX preamble.
     #
     # 'preamble': '',
-
     # Latex figure (float) alignment
     #
     # 'figure_align': 'htbp',
@@ -160,8 +154,7 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'MDV.tex', 'MDV Documentation',
-     'Martin Sergeant', 'manual'),
+    (master_doc, "MDV.tex", "MDV Documentation", "Martin Sergeant", "manual"),
 ]
 
 
@@ -169,10 +162,7 @@ latex_documents = [
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
-man_pages = [
-    (master_doc, 'MDV', 'MDV Documentation',
-     [author], 1)
-]
+man_pages = [(master_doc, "MDV", "MDV Documentation", [author], 1)]
 
 
 # -- Options for Texinfo output ----------------------------------------------
@@ -181,9 +171,15 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'MDV', 'MDV Documentation',
-     author, 'MDV', 'One line description of project.',
-     'Miscellaneous'),
+    (
+        master_doc,
+        "MDV",
+        "MDV Documentation",
+        author,
+        "MDV",
+        "One line description of project.",
+        "Miscellaneous",
+    ),
 ]
 
 
@@ -202,7 +198,7 @@ epub_title = project
 # epub_uid = ''
 
 # A list of files that should not be packed into the epub file.
-epub_exclude_files = ['search.html']
+epub_exclude_files = ["search.html"]
 
 
 # -- Extension configuration -------------------------------------------------
