@@ -2,7 +2,8 @@ import {getRandomString} from "../utilities/Utilities.js";
 import {ContextMenu} from "../utilities/ContextMenu.js";
 import {createEl} from "../utilities/Elements.js";
 import SettingsDialog from "../utilities/SettingsDialog";
-import { chartTypes } from "./ChartTypes.ts";
+import { chartTypes } from "./ChartTypes";
+import DebugJsonDialogReactWrapper from "../react/components/DebugJsonDialogReactWrapper";
 
 
 class BaseChart{
@@ -98,7 +99,10 @@ class BaseChart{
                 menu.push({
                     text:"debug chart",
                     icon:"fas fa-bug",
-                    func:()=>window.mdv.debugChart = this
+                    func:()=>{
+                        window.mdv.debugChart = this;
+                        this.dialogs.push(new DebugJsonDialogReactWrapper(this.config, this));
+                    }
                 });
                 menu.push({
                     text: "copy config JSON to clipboard",
@@ -127,8 +131,14 @@ class BaseChart{
                 if (this.contentDiv !== document.fullscreenElement) console.error('unexpected fullscreen element');
                 const rect = this.contentDiv.getBoundingClientRect();
                 this.setSize(rect.width, rect.height);
+                for (const d of this.dialogs) {
+                    d.setParent(this.contentDiv);
+                }
             } else {
                 this.setSize(...oldSize);
+                for (const d of this.dialogs) {
+                    d.setParent(null);
+                }
             }
         });
         this.addMenuIcon("fas fa-expand","fullscreen", {
@@ -145,7 +155,7 @@ class BaseChart{
         //work out width and height based on container
         this._setDimensions();
     }
-
+    dialogs = [];
     _getContentDimensions(){
         return{ //PJT to review re. gridstack.
             top:5,
@@ -441,8 +451,8 @@ class BaseChart{
         if (this._tooltip){
             this._tooltip.remove();
         }
-        if (this.settingsDialog){
-            this.settingsDialog.close();
+        for (const d of this.dialogs){
+            d.close();
         }
         // dynamic props?
     }
@@ -587,8 +597,12 @@ class BaseChart{
 
    
     _openSettingsDialog(e){
-        if (!this.settingsDialog){
-          this.settingsDialog=  new SettingsDialog({
+        if (!this.settingsDialog) {
+            if (import.meta.env.DEV) {
+                // this.settingsDialog = new SettingsDialogReactWrapper(this);
+                // return;
+            };
+            this.settingsDialog = new SettingsDialog({
                 maxHeight:400,
                 doc:this.__doc__ || document,
                 width:300,
@@ -596,11 +610,13 @@ class BaseChart{
                 position:[e.pageX,e.pageY],
                 useMobx: this.useMobx,
                 onclose:()=>this.settingsDialog=null
-
             },this.getSettings());
-
+            this.dialogs.push(this.settingsDialog);
         }
-      
+        //experimenting with making the parent always be contentDiv (not only in fullscreen mode)
+        //doesn't work ATM because of stacking context - may want to review that more generally 
+        //(can be annoying with tooltips...)
+        //this.settingsDialog.setParent(this.contentDiv);
     }
 
   
@@ -641,6 +657,15 @@ class BaseChart{
         if (this.addToContextMenu){
             menu=menu.concat(this.addToContextMenu());
         }
+        menu.push({
+            text: "experimental settings dialog",
+            icon: "fas fa-cog",
+            func: async () => {
+                const m = await import("../react/components/SettingsDialogReactWrapper");
+                const SettingsDialogReactWrapper = m.default;
+                this.dialogs.push(new SettingsDialogReactWrapper(this));
+            }
+        });
         return menu;
     }
 
@@ -660,8 +685,8 @@ class BaseChart{
     changeBaseDocument(doc){
         this.contextMenu.__doc__=doc;
         this.__doc__=doc;
-        if (this.settingsDialog){
-            this.settingsDialog.close();
+        for (const d of this.dialogs){
+            d.close();
         }
         if (this.legend){
             this.legend.__doc__=doc;
