@@ -1,12 +1,11 @@
-import { PropsWithChildren, createContext, useContext, useMemo, useRef } from "react";
-import { useConfig } from "../../hooks";
-import { ColorPaletteExtension, loadOmeTiff } from "@hms-dbmi/viv";
-import { useChart, useOmeTiff } from "../../context";
+import { PropsWithChildren, createContext, useContext, useRef } from "react";
+import { loadOmeTiff } from "@hms-dbmi/viv";
+import { useChart } from "../../context";
 import { createStore } from "zustand";
 import { useStoreWithEqualityFn } from "zustand/traditional";
 import { observer } from "mobx-react-lite";
 import { VivMDVReact } from "../VivMDVReact";
-import type { ExtractState, ZustandStore } from "./zustandTypes";
+import type { EqFn, Selector, ZustandStore } from "./zustandTypes";
 
 // what about loadOmeZarr, loadBioformatsZarr...
 // ... not to mention HTJ2K?
@@ -249,8 +248,9 @@ export function createVivStores(chart: VivMDVReact) {
         return { ...state, isChannelLoading: newIsChannelLoading };
       })
   })) satisfies VivContextType['viewerStore'];
-  // not sure why 'channelsStore' is misbehaving when other types are ok...
-  return { viewerStore, channelsStore, imageSettingsStore } as unknown as VivContextType;
+  // using `as` instead of `satisfies` because of a bug in the type-checking...
+  // not sure why 'channelsStore' is misbehaving when other types are ok.
+  return { viewerStore, channelsStore, imageSettingsStore } as VivContextType;
 }
 
 const VivContext = createContext<VivContextType>(null);
@@ -265,16 +265,14 @@ const VivContext = createContext<VivContextType>(null);
 export const VivProvider = observer(({ children }: PropsWithChildren) => {
   const vivChart = useChart() as VivMDVReact; //may want to be less MDV-centric here
   if (!vivChart.vivStores) vivChart.vivStores = createVivStores(vivChart); //<< not very react-y: is this ok?
-  const vivStoresRef = useRef(vivChart.vivStores);
   return (
-    <VivContext.Provider value={vivStoresRef.current}>
+    <VivContext.Provider value={vivChart.vivStores}>
       {children}
     </VivContext.Provider>
   )
 });
 
 type StoreName = keyof VivContextType;
-type Selector<S, U> = (state: ExtractState<S>) => U;
 type ImageSettingsStore = VivContextType['imageSettingsStore'];
 type ViewerStore = VivContextType['viewerStore'];
 type ChannelsStore = VivContextType['channelsStore'];
@@ -289,8 +287,6 @@ function useStoreApi<S extends StoreName>(storeName: S): VivContextType[S] {
 export const useChannelsStoreApi = () => useStoreApi('channelsStore');
 export const useImageSettingsStoreApi = () => useStoreApi('imageSettingsStore');
 export const useViewerStoreApi = () => useStoreApi('viewerStore');
-
-type EqFn<U> = (a: U, b: U) => boolean;
 
 /** should be more-or-less equivalent to equivalent avivator hook - 
  * but there can be multiple viv viewers, so we have context for that.
@@ -323,26 +319,3 @@ export const useMetadata = () => {
   return Array.isArray(metadata) ? metadata[image] : metadata;
 };
 
-
-///---------
-
-/** @deprecated this one maybe shouldn't exist */
-export function useChannelsState() {
-  const config = useConfig<any>();
-  //todo something useful... some values...
-  return (config.image_properties || config.viv.image_properties) as ChannelsState;
-}
-
-/** @deprecated this one maybe shouldn't exist */
-export function useVivLayerConfig() {
-  const imageProps = useChannelsState();
-  const extensions = useMemo(() => [new ColorPaletteExtension()], []);
-
-  const ome = useOmeTiff();
-  if (!ome) return undefined;
-  return {
-    ...imageProps,
-    loader: ome.data,
-    extensions
-  }
-}
