@@ -125,6 +125,13 @@ constructor (config={},content) {
     
 
   }
+  setParent(parent) {
+    if (parent) {
+      parent.append(this.outer);
+    } else {
+      this.config.doc.body.append(this.outer);
+    }
+  }
 
   _addFooter(){
     this.footer= createEl("div",{
@@ -203,7 +210,12 @@ constructor (config={},content) {
   * @param {integer} x the new width of the dialog
   * @param {integer} y the new height of the dialog
   */
-  onResize(x,y){}
+  onResize(x, y) {
+    if (this.outer) {
+        this.outer.style.width = x + "px";
+        this.outer.style.height = y + "px";
+    }
+}
 
 
 }
