@@ -8,11 +8,11 @@ import { action } from "mobx";
 class SettingsDialog extends BaseDialog{
     constructor(config,content){
         super(config,content);
-        
     }
-
-    init(content){
-        this.controls=[]
+    
+    init(content, parent){
+        if (!this.controls) this.controls=[];
+        if (!parent) parent = this.dialog;
         //experimental lil-gui version...
         // this.initLilGui(content);
         // return;
@@ -21,7 +21,7 @@ class SettingsDialog extends BaseDialog{
                 styles:{
                     // padding:"5px"
                 }
-            },this.dialog);
+            },parent);
             if (s.type !== "button"){
                 createEl("label",{text:s.label},d);
             }
@@ -40,7 +40,7 @@ class SettingsDialog extends BaseDialog{
                 // we could check whether we'd already wrapped the function, but getSettings() returns a new object each time
                 s.func = action(`${s.label} <action>`, s.func);
             }
-            this.controls[s.label] = this[s.type](s,d);
+            this.controls[s.label] = this[s.type](s,d); //this is going to go wrong if the label isn't unique - like if we have multiple similar layers
         }
     }
     initLilGui(content){
@@ -77,7 +77,7 @@ class SettingsDialog extends BaseDialog{
             
     /** TODO */
     folder(s, d) {
-        const container = d;
+        this.init(s.current_value);
     }
     spinner(s,d){
          
@@ -171,12 +171,14 @@ class SettingsDialog extends BaseDialog{
         }, wrapper);
         createEl("br",{},d);
         for (let item of s.values[0]){
-            const v =item[s.values[2]];
+            // pass 1d array for simple list of strings, or [object[], text_field, value_field] for objects
+            const text = s.values.length > 1 ? item[s.values[1]] : item;
+            const value = s.values.length > 1 ? item[s.values[2]] : item;
             const args = {
-                text:item[s.values[1]],
-                value:v
+                text,
+                value
             }
-            const sel = s.current_value.indexOf(v) !== -1;
+            const sel = s.current_value.indexOf(value) !== -1;
             if (sel){
                 args.selected=true;
             }
@@ -201,8 +203,6 @@ class SettingsDialog extends BaseDialog{
     
 
     dropdown(s, d){
-        //todo fuzzy search / filter / autocomplete
-        //(also for multidropdown)
         const wrapper = createEl("div");
         const dd = createEl("select",{
             styles:{
